msaf/pyproject.toml
[project]
name = "agentic-layer-sdk-msaf"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "agentic-layer-sdk",
    "agent-framework-core>=1.0.0rc2",
    "httpx-retries>=0.4.5",
]

[dependency-groups]
dev = [
    "mypy[reports]>=1.12.0,<2",
    "ruff>=0.11.2",
    "pytest>=9.0.1,<10",
    "pytest-cov>=7,<8",
    "pytest-asyncio>=1.3.0",
    "asgi-lifespan>=2.1.0",
    "respx>=0.22.0",
    "fastmcp>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agenticlayer"]

#[tool.uv]
#prerelease = "allow"

[tool.uv.sources]
agentic-layer-sdk = { path = "../shared" }


[tool.pytest.ini_options]
minversion = "7.0"
testpaths = "tests"
pythonpath = ["."]
addopts = """\
--strict-config \
--cov agenticlayer/msaf \
--cov agenticlayer \
--cov-report html \
--import-mode=importlib \
--show-capture=no \
"""
#filterwarnings = [
#    "ignore::DeprecationWarning:a2a.*",
#    "ignore::UserWarning:agent_framework.*",
#    "ignore::DeprecationWarning:agent_framework.*",
#]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["Q", "N", "I001", "PT", "TID251", "DTZ"]

[tool.mypy]
strict = true
mypy_path = ["."]
namespace_packages = true
explicit_package_bases = true
